Fix 'ArgumentError' is not defined
To my knowledge, ArgumentError is not a standard python exception but comes from the module argparse. argparse.ArgumentError is for program arguments parsing, not for functions/methods arguments bad values.
In addition, the argparse module is not imported, thus when this line is executed, I get the error: "NameError: name 'ArgumentError' is not defined".

That is why I suggest to raise an Exception object.