Prevent margin operation from extending box beyond webmercator bounds.
The `extend_box` function could extend the coordinate box beyond
webmercator bounds (lat: -80, 80, lon: -180, 180). The
tileserver will respond with in-bounds tiles and `imshow` renders
the map visually correct. However, `to_pixels` would calculate coordinates
based on the out-of-bounds box and return values that do not correspond
to the rendered image. Values returned by `extend_box` are now clipped to
bounds so that this cannot happen.